Add comments, implement --force, and ensure the destructor is inside the try block.
--HG--
extra : convert_revision : svn%3A4fd3d1b2-7acc-2142-88d6-33460400c3cb/ladder/trunk%409082

+++ system/core/commands/remove.php
@@ -25,12 +25,33 @@ while ($db->next_database()) {
 			continue;
 		}
 
+		// Initialise the success var.
+		$success = FALSE;
+
+		// Output info.
 		echo 'Downgrading...', "\n";
+
+		// Attempt to run the migration.
 		try {
+			// Run the down method first.
 			$migration->_down();
-			$db->remove_migration($migration_id);
+
+			// Force the destructor to run inside the 'try' block.
+			unset($migration);
+
+			// Success!
+			$success = TRUE;
 		} catch (Exception $e) {
 			echo 'Error: ', $e->getMessage(), "\n";
+		}
+
+		// If it succeeded, or --force is specified, update the migrations.
+		if ($success OR $params['force']) {
+			try {
+				$db->remove_migration($migration_id);
+			} catch (Exception $e) {
+				echo 'Error: ', $e->getMessage(), "\n";
+			}
 		}
 	}
 }
